GA-40 fixed a small bug in Dockerfile, checked that indeed received the server and port in main.py file
